feat(ffe-accordion-react): Make AccordionItems controllable
BREAKING CHANGE: The AccordionItem will now be a controlled component
if the `isOpen` prop is set. This means that it will only open or
close if the `isOpen` prop changes. Previously, the isOpen prop only
controlled the initial isOpen state of an AccordionItem component.

To migrate to this version, check if you're using the `isOpen` prop
anywhere. If you are, please ensure that the `isOpen` prop is controlled
from the consumer's side. You can react to clicks on the accordion item
header with passing a event listener to the `onToggle` prop.

BREAKING CHANGE: The `onOpen` and `onClose` props are removed from the
`AccordionItem` component, and replaced with a similar
`onToggle(nextIsOpen)` prop. `nextIsOpen` will be `true` if the
accordion item _was_ closed, but _will now be_ open, and will be `false`
if the accordion item _was_ open, but _will now be_ closed.

Fixes 155.